server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.4.3
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.13.3
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==9.1.2
tqdm==4.66.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.27.1
websockets==12.0
